frontend/requirements.txt
Django==5.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dateutil==2.8.2
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.9.0
asgiref==3.7.2
six==1.16.0
openpyxl==3.1.2
python-docx==1.1.0
selenium==4.18.1
seleniumbase==4.24.0
